refactor: 💡 refact runtime
